+++ e2e/fixture-package-babel-esm/index.d.ts
@@ -1,7 +1,0 @@
-export default function square(x: number): number;
-export const version: string;
-
-export let counter: number;
-export function countUp(): void;
-
-export function getThis<T>(this: T): T;