+++ packages/babel-plugin-node-cjs-interop/src/.babelrc.js
@@ -1,3 +1,0 @@
-const { createConfig } = require("../configs/babelrc.base.js");
-
-module.exports = createConfig({ esm: process.env.BUILD_TARGET === "esm" });