+++ packages/babel-plugin-node-cjs-interop/test/.babelrc.js
@@ -1,3 +1,0 @@
-const { createConfig } = require("../configs/babelrc.base.js");
-
-module.exports = createConfig({ esm: false });