+++ packages/babel-plugin-node-cjs-interop/test/fixtures/ts-twisted/multiple-import-decls/output.mjs
@@ -1,0 +1,31 @@
+const ns2b = _interopImportCJSNamespaceT(_nsOrig5);
+const ns2 = _interopImportCJSNamespaceT(_nsOrig4);
+const _ns3 = _interopImportCJSNamespaceT(_nsOrig3);
+const _ns2 = _interopImportCJSNamespaceT(_nsOrig2);
+const _ns = _interopImportCJSNamespaceT(_nsOrig);
+function _interopImportCJSNamespaceT(ns) {
+  return ns.default && ns.default.__esModule ? ns : {
+    ...ns,
+    default: ns
+  };
+}
+/*#__CJS__*/
+import * as _nsOrig from "mod1";
+/*#__CJS__*/
+import * as _nsOrig2 from "mod2";
+import "mod1";
+import "mod1";
+/*#__CJS__*/
+import * as _nsOrig3 from "mod1";
+/*#__CJS__*/
+import * as _nsOrig4 from "mod2";
+/*#__CJS__*/
+import * as _nsOrig5 from "mod2";
+console.log({
+  f: _ns.default,
+  x: _ns2.a,
+  b: _ns2.b,
+  f2: _ns3.default,
+  ns2,
+  ns2b
+});