+++ packages/swc-plugin-node-cjs-interop/src/lib.rs
@@ -1,8 +1,3 @@
-// For #[plugin_transform]
-#![allow(clippy::not_unsafe_ptr_arg_deref)]
-// For #[plugin_transform]
-#![allow(clippy::too_many_arguments)]
-
 mod options;
 mod package_name;
 
@@ -12,7 +7,7 @@ use std::mem;
 pub use crate::options::Options as TransformOptions;
 use crate::{options::Options, package_name::get_package_name};
 use swc_core::common::comments::{Comment, CommentKind, Comments};
-use swc_core::common::{Mark, Span, DUMMY_SP};
+use swc_core::common::{Mark, Span, Spanned, SyntaxContext, DUMMY_SP};
 use swc_core::ecma::ast::*;
 use swc_core::ecma::atoms::{js_word, JsWord};
 use swc_core::ecma::visit::{as_folder, noop_visit_mut_type, FoldWith, VisitMut, VisitMutWith};
@@ -45,6 +40,7 @@ struct ModuleVisitor<'a, C: Comments> {
     parent: &'a TransformVisitor<C>,
     replace_map: HashMap<Id, PseudoImport>,
     import_helper_name: Option<Ident>,
+    import_helper_name_t: Option<Ident>,
     prepend_body: Vec<ModuleItem>,
 }
 
@@ -54,6 +50,7 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
             parent,
             replace_map: HashMap::new(),
             import_helper_name: None,
+            import_helper_name_t: None,
             prepend_body: Vec::new(),
         }
     }
@@ -62,6 +59,7 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
         for stmt in &mut module.body {
             self.process_module_item(stmt);
         }
+        self.visit_mut_module(module);
         if !self.replace_map.is_empty() {
             module.visit_mut_with(&mut ModuleRefRewriter {
                 replace_map: &self.replace_map,
@@ -74,9 +72,9 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
 
     fn process_module_item(&mut self, stmt: &mut ModuleItem) {
         if let ModuleItem::ModuleDecl(ModuleDecl::Import(stmt)) = stmt {
-            if !self.parent.options.has_applicable_source(&stmt.src.value) {
+            let Some(pkg_type) = self.parent.options.applicable_source_type(&stmt.src.value) else {
                 return;
-            }
+            };
             // Should have been removed by transform-typescript. Just in case.
             if stmt.type_only {
                 return;
@@ -88,7 +86,7 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                 return;
             }
 
-            let macro_span = DUMMY_SP.apply_mark(Mark::new());
+            let macro_ctxt = SyntaxContext::empty().apply_mark(Mark::new());
             let existing_ns_import = stmt
                 .specifiers
                 .iter()
@@ -97,7 +95,7 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
             let ns_import = if let Some(ns_import) = existing_ns_import {
                 ns_import
             } else {
-                Ident::new(JsWord::from("_ns"), macro_span)
+                Ident::new(JsWord::from("_ns"), DUMMY_SP, macro_ctxt)
             };
 
             for specifier in &stmt.specifiers {
@@ -105,7 +103,7 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                     // ns.default
                     ImportSpecifier::Default(_) => PseudoImport::Member(
                         ns_import.clone(),
-                        Ident::new(js_word!("default"), DUMMY_SP),
+                        IdentName::new(js_word!("default"), DUMMY_SP),
                     ),
                     ImportSpecifier::Named(specifier) => match &specifier.imported {
                         // ns["named"]
@@ -114,10 +112,12 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                         }
                         // ns.named
                         Some(ModuleExportName::Ident(imported)) => {
-                            PseudoImport::Member(ns_import.clone(), imported.clone())
+                            PseudoImport::Member(ns_import.clone(), imported.clone().into())
                         }
                         // ns.named
-                        None => PseudoImport::Member(ns_import.clone(), specifier.local.clone()),
+                        None => {
+                            PseudoImport::Member(ns_import.clone(), specifier.local.clone().into())
+                        }
                     },
                     // No need to replace
                     ImportSpecifier::Namespace(_) => continue,
@@ -130,16 +130,17 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                 self.replace_map.insert(local_id, expr);
             }
 
-            let import_helper = self.get_import_helper();
+            let import_helper = self.get_import_helper(pkg_type);
 
             // import ... from "source";
             // ->
             // import * as moduleOrig from "source";
             // const module = _interopImportCJSNamespace(moduleOrig);
-            let import_original_name = Ident::new(JsWord::from("_nsOrig"), macro_span);
+            let import_original_name = Ident::new(JsWord::from("_nsOrig"), DUMMY_SP, macro_ctxt);
             self.prepend_body.push(
                 Stmt::Decl(Decl::Var(Box::new(VarDecl {
                     span: Span::dummy_with_cmt(),
+                    ctxt: SyntaxContext::empty(),
                     kind: VarDeclKind::Const,
                     declare: false,
                     decls: vec![VarDeclarator {
@@ -148,6 +149,7 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                         init: Some(Box::new(
                             CallExpr {
                                 span: DUMMY_SP,
+                                ctxt: SyntaxContext::empty(),
                                 callee: Box::new(Expr::Ident(import_helper)).into(),
                                 args: if self.parent.options.loose {
                                     vec![
@@ -176,13 +178,122 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
         }
     }
 
-    fn get_import_helper(&mut self) -> Ident {
-        if let Some(helper) = &self.import_helper_name {
+    fn process_dynamic_import(&mut self, mut call: ImportExprCtx<'_>) {
+        let Some(source) = call.as_call().args[0].expr.as_lit().and_then(|lit| {
+            if let Lit::Str(s) = lit {
+                Some(s)
+            } else {
+                None
+            }
+        }) else {
+            return;
+        };
+        let Some(pkg_type) = self.parent.options.applicable_source_type(&source.value) else {
+            return;
+        };
+        let import_helper = self.get_import_helper(pkg_type);
+
+        if is_cjs_annotated(call.as_call_wrapper(), &self.parent.comments) {
+            return;
+        }
+        match &mut call {
+            ImportExprCtx::AwaitImport(awaiter_path) => {
+                if is_cjs_annotated(*awaiter_path, &self.parent.comments) {
+                    return;
+                }
+                // await import("source")
+                // -> _interopImportCJSNamespace(await import("source"))
+                let awaiter =
+                    mem::replace(*awaiter_path, Expr::Invalid(Invalid { span: DUMMY_SP }));
+                annotate_as_cjs(&awaiter, &self.parent.comments);
+                **awaiter_path = Expr::Call(CallExpr {
+                    span: DUMMY_SP,
+                    ctxt: SyntaxContext::empty(),
+                    callee: Callee::Expr(Box::new(Expr::Ident(import_helper.clone()))),
+                    args: if pkg_type != PackageType::TsTwisted && self.parent.options.loose {
+                        vec![awaiter.into(), Expr::Lit(Lit::Bool(true.into())).into()]
+                    } else {
+                        vec![awaiter.into()]
+                    },
+                    type_args: None,
+                });
+            }
+            ImportExprCtx::Import(call_path) => {
+                // import("source")
+                // -> import("source").then((ns) => _interopImportCJSNamespace(ns))
+                // NOTE: strictly speaking, it does not preserve scheduling semantics of the Promise;
+                //       the transform delays the execution by one cycle of microtask queue.
+                //       We could add other heuristics for cases like import("source").then(...)
+                //       but ultimately we cannot do so when we are not sure where the Promise goes to.
+                //       As the effect is fairly minor (do not depend on microtask cycle count!),
+                //       we just leave it as is for now.
+                let ns = Ident::new_no_ctxt(JsWord::from("ns"), DUMMY_SP);
+                let call = mem::replace(*call_path, Expr::Invalid(Invalid { span: DUMMY_SP }));
+                annotate_as_cjs(&call, &self.parent.comments);
+                **call_path = Expr::Call(CallExpr {
+                    span: DUMMY_SP,
+                    ctxt: SyntaxContext::empty(),
+                    callee: Callee::Expr(Box::new(Expr::Member(MemberExpr {
+                        span: DUMMY_SP,
+                        // Explicit parentheses to get room for the #__CJS__ comment -- but seems not working
+                        obj: Box::new(Expr::Paren(ParenExpr {
+                            span: DUMMY_SP,
+                            expr: Box::new(call),
+                        })),
+                        prop: MemberProp::Ident(IdentName::new(JsWord::from("then"), DUMMY_SP)),
+                    }))),
+                    args: vec![Expr::Arrow(ArrowExpr {
+                        span: DUMMY_SP,
+                        ctxt: SyntaxContext::empty(),
+                        params: vec![Pat::Ident(ns.clone().into())],
+                        body: Box::new(BlockStmtOrExpr::Expr(Box::new(
+                            CallExpr {
+                                span: DUMMY_SP,
+                                ctxt: SyntaxContext::empty(),
+                                callee: Callee::Expr(Box::new(Expr::Ident(import_helper.clone()))),
+                                args: if pkg_type != PackageType::TsTwisted
+                                    && self.parent.options.loose
+                                {
+                                    vec![
+                                        Expr::Ident(ns.clone()).into(),
+                                        Expr::Lit(Lit::Bool(true.into())).into(),
+                                    ]
+                                } else {
+                                    vec![Expr::Ident(ns.clone()).into()]
+                                },
+                                type_args: None,
+                            }
+                            .into(),
+                        ))),
+                        is_async: false,
+                        is_generator: false,
+                        type_params: None,
+                        return_type: None,
+                    })
+                    .into()],
+                    type_args: None,
+                });
+            }
+        }
+    }
+
+    fn get_import_helper(&mut self, pkg_type: PackageType) -> Ident {
+        if let Some(helper) = match pkg_type {
+            PackageType::Normal => &self.import_helper_name,
+            PackageType::TsTwisted => &self.import_helper_name_t,
+        } {
             return helper.clone();
         }
 
-        let macro_span = DUMMY_SP.apply_mark(Mark::new());
-        let helper = { Ident::new(JsWord::from("_interopImportCJSNamespace"), macro_span) };
+        let macro_ctxt = SyntaxContext::empty().apply_mark(Mark::new());
+        let helper = Ident::new(
+            JsWord::from(match pkg_type {
+                PackageType::Normal => "_interopImportCJSNamespace",
+                PackageType::TsTwisted => "_interopImportCJSNamespaceT",
+            }),
+            DUMMY_SP,
+            macro_ctxt,
+        );
 
         if self.parent.options.use_runtime {
             // import {
@@ -195,7 +306,8 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                         span: DUMMY_SP,
                         local: helper.clone(),
                         imported: Some(
-                            Ident::new(JsWord::from("interopImportCJSNamespace"), DUMMY_SP).into(),
+                            Ident::new_no_ctxt(JsWord::from("interopImportCJSNamespace"), DUMMY_SP)
+                                .into(),
                         ),
                         is_type_only: false,
                     }
@@ -203,115 +315,266 @@ impl<'a, C: Comments> ModuleVisitor<'a, C> {
                     src: Box::new(Str::from("node-cjs-interop")),
                     type_only: false,
                     with: None,
+                    phase: ImportPhase::Evaluation,
                 })
                 .into(),
             );
 
-            self.import_helper_name = Some(helper.clone());
+            *match pkg_type {
+                PackageType::Normal => &mut self.import_helper_name,
+                PackageType::TsTwisted => &mut self.import_helper_name_t,
+            } = Some(helper.clone());
             return helper;
         }
 
-        let fn_span = DUMMY_SP.apply_mark(Mark::new());
-        let ns = Ident::new(JsWord::from("ns"), fn_span);
-        let loose = Ident::new(JsWord::from("loose"), fn_span);
+        let fn_ctxt = SyntaxContext::empty().apply_mark(Mark::new());
+        let ns = Ident::new(JsWord::from("ns"), DUMMY_SP, fn_ctxt);
+        let loose = Ident::new(JsWord::from("loose"), DUMMY_SP, fn_ctxt);
         let ns_default = Expr::Member(MemberExpr {
             span: DUMMY_SP,
             obj: Box::new(ns.clone().into()),
-            prop: Ident::new(js_word!("default"), DUMMY_SP).into(),
+            prop: IdentName::new(js_word!("default"), DUMMY_SP).into(),
         });
-        // function interopImportCJSNamespace(ns) {
-        //   return ns.__esModule && ns.default && ns.default.__esModule ? ns.default : ns;
-        // }
-        self.prepend_body.push(
-            Stmt::Decl(Decl::Fn(FnDecl {
-                ident: helper.clone(),
-                declare: false,
-                function: Box::new(Function {
-                    params: vec![ns.clone().into(), loose.clone().into()],
-                    decorators: vec![],
-                    span: Span::dummy_with_cmt(),
-                    body: Some(BlockStmt {
-                        span: DUMMY_SP,
-                        stmts: vec![ReturnStmt {
-                            span: DUMMY_SP,
-                            arg: Some(Box::new(
-                                CondExpr {
+        match pkg_type {
+            PackageType::TsTwisted => {
+                // function interopImportCJSNamespaceT(ns) {
+                //   return ns.default && ns.default.__esModule ? ns : {
+                //     ...ns,
+                //     default: ns
+                //   };
+                // }
+                self.prepend_body.push(
+                    Stmt::Decl(Decl::Fn(FnDecl {
+                        ident: helper.clone(),
+                        declare: false,
+                        function: Box::new(Function {
+                            params: vec![ns.clone().into()],
+                            decorators: vec![],
+                            span: Span::dummy_with_cmt(),
+                            ctxt: SyntaxContext::empty(),
+                            body: Some(BlockStmt {
+                                span: DUMMY_SP,
+                                ctxt: SyntaxContext::empty(),
+                                stmts: vec![ReturnStmt {
                                     span: DUMMY_SP,
-                                    test: Box::new(
-                                        BinExpr {
+                                    arg: Some(Box::new(
+                                        CondExpr {
                                             span: DUMMY_SP,
-                                            op: BinaryOp::LogicalAnd,
-                                            left: Box::new(
+                                            test: Box::new(
+                                                BinExpr {
+                                                    span: DUMMY_SP,
+                                                    op: BinaryOp::LogicalAnd,
+                                                    left: Box::new(ns_default.clone()),
+                                                    right: Box::new(
+                                                        MemberExpr {
+                                                            span: DUMMY_SP,
+                                                            obj: Box::new(ns_default.clone()),
+                                                            prop: IdentName::new(
+                                                                JsWord::from("__esModule"),
+                                                                DUMMY_SP,
+                                                            )
+                                                            .into(),
+                                                        }
+                                                        .into(),
+                                                    ),
+                                                }
+                                                .into(),
+                                            ),
+                                            cons: Box::new(ns.clone().into()),
+                                            alt: Box::new(
+                                                ObjectLit {
+                                                    span: DUMMY_SP,
+                                                    props: vec![
+                                                        PropOrSpread::Spread(SpreadElement {
+                                                            dot3_token: DUMMY_SP,
+                                                            expr: Box::new(ns.clone().into()),
+                                                        }),
+                                                        PropOrSpread::Prop(Box::new(
+                                                            KeyValueProp {
+                                                                key: PropName::Ident(
+                                                                    IdentName::new(
+                                                                        JsWord::from("default"),
+                                                                        DUMMY_SP,
+                                                                    ),
+                                                                ),
+                                                                value: Box::new(ns.clone().into()),
+                                                            }
+                                                            .into(),
+                                                        )),
+                                                    ],
+                                                }
+                                                .into(),
+                                            ),
+                                        }
+                                        .into(),
+                                    )),
+                                }
+                                .into()],
+                            }),
+                            is_generator: false,
+                            is_async: false,
+                            type_params: None,
+                            return_type: None,
+                        }),
+                    }))
+                    .into(),
+                );
+            }
+            PackageType::Normal => {
+                // function interopImportCJSNamespace(ns, loose) {
+                //   return (loose || ns.__esModule) && ns.default && ns.default.__esModule ? ns.default : ns;
+                // }
+                self.prepend_body.push(
+                    Stmt::Decl(Decl::Fn(FnDecl {
+                        ident: helper.clone(),
+                        declare: false,
+                        function: Box::new(Function {
+                            params: vec![ns.clone().into(), loose.clone().into()],
+                            decorators: vec![],
+                            span: Span::dummy_with_cmt(),
+                            ctxt: SyntaxContext::empty(),
+                            body: Some(BlockStmt {
+                                span: DUMMY_SP,
+                                ctxt: SyntaxContext::empty(),
+                                stmts: vec![ReturnStmt {
+                                    span: DUMMY_SP,
+                                    arg: Some(Box::new(
+                                        CondExpr {
+                                            span: DUMMY_SP,
+                                            test: Box::new(
                                                 BinExpr {
                                                     span: DUMMY_SP,
                                                     op: BinaryOp::LogicalAnd,
                                                     left: Box::new(
                                                         BinExpr {
                                                             span: DUMMY_SP,
-                                                            op: BinaryOp::LogicalOr,
-                                                            left: Box::new(loose.clone().into()),
-                                                            right: Box::new(
-                                                                MemberExpr {
+                                                            op: BinaryOp::LogicalAnd,
+                                                            left: Box::new(
+                                                                BinExpr {
                                                                     span: DUMMY_SP,
-                                                                    obj: Box::new(
-                                                                        ns.clone().into(),
+                                                                    op: BinaryOp::LogicalOr,
+                                                                    left: Box::new(
+                                                                        loose.clone().into(),
                                                                     ),
-                                                                    prop: Ident::new(
-                                                                        JsWord::from("__esModule"),
-                                                                        DUMMY_SP,
-                                                                    )
-                                                                    .into(),
+                                                                    right: Box::new(
+                                                                        MemberExpr {
+                                                                            span: DUMMY_SP,
+                                                                            obj: Box::new(
+                                                                                ns.clone().into(),
+                                                                            ),
+                                                                            prop: IdentName::new(
+                                                                                JsWord::from(
+                                                                                    "__esModule",
+                                                                                ),
+                                                                                DUMMY_SP,
+                                                                            )
+                                                                            .into(),
+                                                                        }
+                                                                        .into(),
+                                                                    ),
                                                                 }
                                                                 .into(),
                                                             ),
+                                                            right: Box::new(ns_default.clone()),
                                                         }
                                                         .into(),
                                                     ),
-                                                    right: Box::new(ns_default.clone()),
+                                                    right: Box::new(
+                                                        MemberExpr {
+                                                            span: DUMMY_SP,
+                                                            obj: Box::new(ns_default.clone()),
+                                                            prop: IdentName::new(
+                                                                JsWord::from("__esModule"),
+                                                                DUMMY_SP,
+                                                            )
+                                                            .into(),
+                                                        }
+                                                        .into(),
+                                                    ),
                                                 }
                                                 .into(),
                                             ),
-                                            right: Box::new(
-                                                MemberExpr {
-                                                    span: DUMMY_SP,
-                                                    obj: Box::new(ns_default.clone()),
-                                                    prop: Ident::new(
-                                                        JsWord::from("__esModule"),
-                                                        DUMMY_SP,
-                                                    )
-                                                    .into(),
-                                                }
-                                                .into(),
-                                            ),
+                                            cons: Box::new(ns_default),
+                                            alt: Box::new(ns.into()),
                                         }
                                         .into(),
-                                    ),
-                                    cons: Box::new(ns_default),
-                                    alt: Box::new(ns.into()),
+                                    )),
                                 }
-                                .into(),
-                            )),
-                        }
-                        .into()],
-                    }),
-                    is_generator: false,
-                    is_async: false,
-                    type_params: None,
-                    return_type: None,
-                }),
-            }))
-            .into(),
-        );
+                                .into()],
+                            }),
+                            is_generator: false,
+                            is_async: false,
+                            type_params: None,
+                            return_type: None,
+                        }),
+                    }))
+                    .into(),
+                );
+            }
+        }
         self.import_helper_name = Some(helper.clone());
         helper
     }
+}
+
+enum ImportExprCtx<'a> {
+    Import(&'a mut Expr),
+    AwaitImport(&'a mut Expr),
+}
+
+impl<'a> ImportExprCtx<'a> {
+    fn as_call(&self) -> &CallExpr {
+        self.as_call_wrapper().as_call().unwrap()
+    }
+    fn as_call_wrapper(&self) -> &Expr {
+        match self {
+            ImportExprCtx::Import(call) => call,
+            ImportExprCtx::AwaitImport(awaiter) => &awaiter.as_await_expr().unwrap().arg,
+        }
+    }
+    // fn as_mut_call(&mut self) -> &mut CallExpr {
+    //     self.as_mut_call_wrapper().as_mut_call().unwrap()
+    // }
+    // fn as_mut_call_wrapper(&mut self) -> &mut Expr {
+    //     match self {
+    //         ImportExprCtx::Import(call) => call,
+    //         ImportExprCtx::AwaitImport(awaiter) => &mut awaiter.as_mut_await_expr().unwrap().arg,
+    //     }
+    // }
+}
+
+impl<'a, C: Comments> VisitMut for ModuleVisitor<'a, C> {
+    noop_visit_mut_type!();
+
+    fn visit_mut_expr(&mut self, n: &mut Expr) {
+        if let Some(awaiter) = n.as_mut_await_expr() {
+            if let Some(call) = awaiter.arg.as_mut_call() {
+                if is_import_expr(call) {
+                    // Skip processing the call expression
+                    call.visit_mut_children_with(self);
+                    self.process_dynamic_import(ImportExprCtx::AwaitImport(n));
+                    return;
+                }
+            }
+        }
+        n.visit_mut_children_with(self);
+        if let Some(call) = n.as_mut_call() {
+            if is_import_expr(call) {
+                self.process_dynamic_import(ImportExprCtx::Import(n));
+            }
+        }
+    }
+}
+
+fn is_import_expr(expr: &CallExpr) -> bool {
+    expr.callee.is_import() && !expr.args.is_empty() && expr.args[0].spread.is_none()
 }
 
 #[derive(Debug, Clone)]
 enum PseudoImport {
     #[allow(dead_code)]
     Ident(Ident),
-    Member(Ident, Ident),
+    Member(Ident, IdentName),
     Member2(Ident, Str),
 }
 
@@ -359,6 +622,7 @@ impl PseudoImport {
         match self {
             PseudoImport::Ident(x) => x.clone().into(),
             PseudoImport::Member(obj, prop) => Box::new(JSXMemberExpr {
+                span: DUMMY_SP,
                 obj: obj.clone().into(),
                 prop: prop.clone(),
             })
@@ -373,6 +637,7 @@ impl PseudoImport {
         match self {
             PseudoImport::Ident(x) => x.clone().into(),
             PseudoImport::Member(obj, prop) => JSXMemberExpr {
+                span: DUMMY_SP,
                 obj: obj.clone().into(),
                 prop: prop.clone(),
             }
@@ -490,8 +755,8 @@ impl<'a> ModuleRefRewriter<'a> {
     }
 }
 
-fn is_cjs_annotated<C: Comments>(stmt: &ImportDecl, comments: &C) -> bool {
-    let leading_comments = comments.get_leading(stmt.span.lo());
+fn is_cjs_annotated<N: Spanned, C: Comments>(node: &N, comments: &C) -> bool {
+    let leading_comments = comments.get_leading(node.span().lo());
     if let Some(leading_comments) = leading_comments {
         leading_comments
             .iter()
@@ -501,10 +766,10 @@ fn is_cjs_annotated<C: Comments>(stmt: &ImportDecl, comments: &C) -> bool {
     }
 }
 
-fn annotate_as_cjs<C: Comments>(stmt: &ImportDecl, comments: &C) {
-    if !is_cjs_annotated(stmt, comments) {
+fn annotate_as_cjs<N: Spanned, C: Comments>(node: &N, comments: &C) {
+    if !is_cjs_annotated(node, comments) {
         comments.add_leading(
-            stmt.span.lo(),
+            node.span().lo(),
             Comment {
                 kind: CommentKind::Block,
                 span: DUMMY_SP,
@@ -514,13 +779,25 @@ fn annotate_as_cjs<C: Comments>(stmt: &ImportDecl, comments: &C) {
     }
 }
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+enum PackageType {
+    Normal,
+    TsTwisted,
+}
+
 impl Options {
-    fn has_applicable_source(&self, source: &str) -> bool {
+    fn applicable_source_type(&self, source: &str) -> Option<PackageType> {
         let source_package = get_package_name(source);
         if let Some(source_package) = source_package {
-            self.packages.iter().any(|pkg| pkg == source_package)
+            if self.packages.iter().any(|pkg| pkg == source_package) {
+                Some(PackageType::Normal)
+            } else if self.packages_t.iter().any(|pkg| pkg == source_package) {
+                Some(PackageType::TsTwisted)
+            } else {
+                None
+            }
         } else {
-            false
+            None
         }
     }
 }