+++ packages/swc-plugin-node-cjs-interop/tests/fixtures/ts-twisted/multiple-imports/output.mjs
@@ -1,0 +1,13 @@
+/*#__CJS__*/ const _ns = _interopImportCJSNamespaceT(_nsOrig);
+function _interopImportCJSNamespaceT(ns) {
+    return ns.default && ns.default.__esModule ? ns : {
+        ...ns,
+        default: ns
+    };
+}
+import * as _nsOrig from "mod";
+console.log({
+    f: _ns.default,
+    x: _ns.a,
+    b: _ns.b
+});