+++ packages/swc-plugin-node-cjs-interop/tests/fixtures/ts-twisted/namespace-imports/output.mjs
@@ -1,0 +1,9 @@
+/*#__CJS__*/ const M = _interopImportCJSNamespaceT(_nsOrig);
+function _interopImportCJSNamespaceT(ns) {
+    return ns.default && ns.default.__esModule ? ns : {
+        ...ns,
+        default: ns
+    };
+}
+import * as _nsOrig from "mod";
+console.log(M);