+++ packages/node-cjs-interop/CHANGELOG.md
@@ -1,4 +1,42 @@
-## Unreleased
+# node-cjs-interop
+
+## 0.1.7
+
+### Patch Changes
+
+- c75999a: chore(deps): update babel monorepo
+
+## 0.1.6
+
+### Patch Changes
+
+- 95bef4e: Update typescript
+- a6b57a9: Update Babel
+
+## 0.1.5
+
+### Patch Changes
+
+- fb98f16: Internally improve Babel config
+
+## 0.1.4
+
+### Patch Changes
+
+- 972e632: Add a variant helper that wraps default one more time.
+
+  The helper is called `interopImportCJSNamespaceT` and you can use it just like `interopImportCJSNamespace`
+  (except that it assumes `loose = true` by default).
+
+  To use it from the Babel/SWC plugins, use the `packagesT` option instead of `packages`.
+
+## 0.1.3
+
+### Patch Changes
+
+- 54b33bb: Change how dual packages are configured
+
+## 0.1.2
 
 - Update dependencies