+++ packages/swc-plugin-node-cjs-interop/CHANGELOG.md
@@ -1,4 +1,85 @@
-## Unreleased
+# swc-plugin-node-cjs-interop
+
+## 0.1.14
+
+### Patch Changes
+
+- 5c6d95a: Update swc monorepo
+
+## 0.1.13
+
+### Patch Changes
+
+- cdcee32: chore(deps): update swc monorepo
+
+## 0.1.12
+
+### Patch Changes
+
+- 392044a: chore(deps): update swc monorepo
+
+## 0.1.11
+
+### Patch Changes
+
+- e788916: chore(deps): update swc monorepo
+
+## 0.1.10
+
+### Patch Changes
+
+- a5beaff: chore(deps): update swc monorepo
+
+## 0.1.9
+
+### Patch Changes
+
+- 5c4ec19: Update swc monorepo
+
+## 0.1.8
+
+### Patch Changes
+
+- be3d9ec: Update swc monorepo
+- a696577: Update serde
+- bcdc489: Update serde_json
+
+## 0.1.7
+
+### Patch Changes
+
+- 9437e69: Bump swc_core to 0.87
+
+## 0.1.6
+
+### Patch Changes
+
+- de38484: Update swc monorepo
+
+## 0.1.5
+
+### Patch Changes
+
+- 37a65b0: Support dynamic imports
+
+## 0.1.4
+
+### Patch Changes
+
+- b16c3ee: Update swc monorepo
+
+## 0.1.3
+
+### Patch Changes
+
+- 972e632: Add a variant helper that wraps default one more time.
+
+  The helper is called `interopImportCJSNamespaceT` and you can use it just like `interopImportCJSNamespace`
+  (except that it assumes `loose = true` by default).
+
+  To use it from the Babel/SWC plugins, use the `packagesT` option instead of `packages`.
+
+## 0.1.2
 
 - Update dependencies